+++ packages/studio/src/features/docker-manager/api/mutations/use-seed-database.ts
@@ -5,7 +5,7 @@ type SeedDatabaseOptions = {
 	containerId: string
 	filePath: string
 	connectionConfig: {
-		provider?: 'postgres' | 'mariadb' | 'cockroach'
+		provider?: 'postgres' | 'mariadb' | 'mysql' | 'cockroach'
 		user: string
 		password?: string
 		database: string